feat(cpu): allow fractional CPU values for sandbox and MicroVm configurations

+++ microsandbox-core/lib/management/home.rs
@@ -217,7 +217,7 @@ pub async fn install(
         let mut b = Sandbox::builder().image(ReferenceOrPath::Reference(image.clone()));
 
         if let Some(cpus) = cpus {
-            b = b.cpus(cpus);
+            b = b.cpus(cpus as f32);
         }
 
         if let Some(memory) = memory {
